move even more code out of the workflow package
